+++ Database/models/AutonomousDatabaseDataguardAssociation.cs
@@ -46,7 +46,8 @@ namespace Oci.DatabaseService.Models
         public string AutonomousDatabaseId { get; set; }
                 ///
         /// <value>
-        /// The role of the Autonomous Data Guard-enabled Autonomous Container Database.
+        /// The Data Guard role of the Autonomous Container Database, if Autonomous Data Guard is enabled.
+        /// 
         /// </value>
         ///
         public enum RoleEnum {
@@ -59,7 +60,8 @@ namespace Oci.DatabaseService.Models
         };
 
         /// <value>
-        /// The role of the Autonomous Data Guard-enabled Autonomous Container Database.
+        /// The Data Guard role of the Autonomous Container Database, if Autonomous Data Guard is enabled.
+        /// 
         /// </value>
         /// <remarks>
         /// Required
@@ -109,7 +111,8 @@ namespace Oci.DatabaseService.Models
         public string LifecycleDetails { get; set; }
                 ///
         /// <value>
-        /// The role of the Autonomous Data Guard-enabled Autonomous Container Database.
+        /// The Data Guard role of the Autonomous Container Database, if Autonomous Data Guard is enabled.
+        /// 
         /// </value>
         ///
         public enum PeerRoleEnum {
@@ -122,7 +125,8 @@ namespace Oci.DatabaseService.Models
         };
 
         /// <value>
-        /// The role of the Autonomous Data Guard-enabled Autonomous Container Database.
+        /// The Data Guard role of the Autonomous Container Database, if Autonomous Data Guard is enabled.
+        /// 
         /// </value>
         /// <remarks>
         /// Required