+++ Database/models/DatabaseSummary.cs
@@ -87,7 +87,7 @@ namespace Oci.DatabaseService.Models
         public string DbName { get; set; }
         
         /// <value>
-        /// The name of the pluggable database. The name must begin with an alphabetic character and can contain a maximum of eight alphanumeric characters. Special characters are not permitted. Pluggable database should not be same as database name.
+        /// The name of the pluggable database. The name must begin with an alphabetic character and can contain a maximum of thirty alphanumeric characters. Special characters are not permitted. Pluggable database should not be same as database name.
         /// </value>
         [JsonProperty(PropertyName = "pdbName")]
         public string PdbName { get; set; }