+++ Datacatalog/models/Attribute.cs
@@ -42,6 +42,12 @@ namespace Oci.DatacatalogService.Models
         public string DisplayName { get; set; }
         
         /// <value>
+        /// Optional user friendly business name of the attribute. If set, this supplements the harvested display name of the object.
+        /// </value>
+        [JsonProperty(PropertyName = "businessName")]
+        public string BusinessName { get; set; }
+        
+        /// <value>
         /// Detailed description of the attribute.
         /// </value>
         [JsonProperty(PropertyName = "description")]