+++ Datacatalog/models/CreateCustomPropertyDetails.cs
@@ -77,6 +77,12 @@ namespace Oci.DatacatalogService.Models
         public System.Nullable<bool> IsEditable { get; set; }
         
         /// <value>
+        /// If this field is displayed in a list view of applicable objects.
+        /// </value>
+        [JsonProperty(PropertyName = "isShownInList")]
+        public System.Nullable<bool> IsShownInList { get; set; }
+        
+        /// <value>
         /// If this field is allowed to pop in search results
         /// </value>
         [JsonProperty(PropertyName = "isHiddenInSearch")]