+++ Datacatalog/models/CustomPropertyGetUsage.cs
@@ -83,6 +83,12 @@ namespace Oci.DatacatalogService.Models
         public System.Nullable<bool> IsEditable { get; set; }
         
         /// <value>
+        /// If this field is displayed in a list view of applicable objects.
+        /// </value>
+        [JsonProperty(PropertyName = "isShownInList")]
+        public System.Nullable<bool> IsShownInList { get; set; }
+        
+        /// <value>
         /// Is this property allowed to have list of values
         /// </value>
         [JsonProperty(PropertyName = "isListType")]